improved logging system

+++ src/PID.cpp
@@ -2,29 +2,29 @@
 #include "config.h"
 #include <Arduino.h>
 
-const double e_rpm_const = 60000000.0 / ENGINE_NUM_MAGS; // 60s/min * 1000ms/s / num_mags
-const double s_rpm_const = 60000000.0 / SECONDARY_NUM_MAGS;
+const float e_rpm_const = 60000000.0 / ENGINE_NUM_MAGS; // 60s/min * 1000ms/s / num_mags
+const float s_rpm_const = 60000000.0 / SECONDARY_NUM_MAGS;
 
 
 // Exported from Octave, currently unused
-const double model_r[] = {0.9, 1, 1.1, 1.2, 1.3, 1.4, 1.5, 1.6, 1.7, 1.8, 1.9, 2, 2.1, 2.2, 2.3, 2.4, 2.5, 2.6, 2.7, 2.8, 2.9, 3, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6, 3.7, 3.8, 3.9};
-const double model_P[] = {1.0007, 0.9295, 0.86404, 0.80379, 0.74825, 0.697, 0.6496, 0.6057, 0.56497, 0.5271, 0.49182, 0.4589, 0.42813, 0.39931, 0.37227, 0.34686, 0.32295, 0.30041, 0.27913, 0.25901, 0.23996, 0.22191, 0.20477, 0.18849, 0.17299, 0.15824, 0.14417, 0.13073, 0.1179, 0.10563, 0.093882};
-double error = 0.0;
-double last_error = 0.0;
-double error_deriv = 0.0;
-double error_integ = 0.0;
-const double max_pos_inch = 0.925 * 2; // *2 from yoke leverage
-double max_error_integ = max_pos_inch / KI_DEFAULT;
-double min_error_integ = -max_error_integ;
-double e_rpm_t = E_RPM_TARGET_DEFAULT;
-double e_rpm_m = 0.0;
-double s_rpm_m = 0.0;
-double r_t = 3.9;
-double r_m = 3.9;
-double target_pos_inch = 0.0;
-double Ki = KI_DEFAULT;
-double Kp = KP_DEFAULT;
-double Kd = KD_DEFAULT;
+const float model_r[] = {0.9, 1, 1.1, 1.2, 1.3, 1.4, 1.5, 1.6, 1.7, 1.8, 1.9, 2, 2.1, 2.2, 2.3, 2.4, 2.5, 2.6, 2.7, 2.8, 2.9, 3, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6, 3.7, 3.8, 3.9};
+const float model_P[] = {1.0007, 0.9295, 0.86404, 0.80379, 0.74825, 0.697, 0.6496, 0.6057, 0.56497, 0.5271, 0.49182, 0.4589, 0.42813, 0.39931, 0.37227, 0.34686, 0.32295, 0.30041, 0.27913, 0.25901, 0.23996, 0.22191, 0.20477, 0.18849, 0.17299, 0.15824, 0.14417, 0.13073, 0.1179, 0.10563, 0.093882};
+float error = 0.0;
+float last_error = 0.0;
+float error_deriv = 0.0;
+float error_integ = 0.0;
+const float max_pos_inch = 0.925 * 2; // *2 from yoke leverage
+float max_error_integ = max_pos_inch / KI_DEFAULT;
+float min_error_integ = -max_error_integ;
+float e_rpm_t = E_RPM_TARGET_DEFAULT;
+float e_rpm_m = 0.0;
+float s_rpm_m = 0.0;
+float r_t = 3.9;
+float r_m = 3.9;
+float target_pos_inch = 0.0;
+float Ki = KI_DEFAULT;
+float Kp = KP_DEFAULT;
+float Kd = KD_DEFAULT;
 
 // Stepper Drive
 FastAccelStepperEngine engine = FastAccelStepperEngine();
@@ -40,7 +40,7 @@ void stepper_setup() {
 }
 
 const unsigned int delta_t = PID_TASK_DELAY; // in ms
-const double dt = delta_t * 0.001; // in seconds for PID
+const float dt = delta_t * 0.001; // in seconds for PID
 
 volatile unsigned long e_new_pulse = 0;
 volatile unsigned long e_last_pulse = 0;
@@ -64,8 +64,8 @@ void e_isr() { // 4 micros
     else e_delta_i++;
 }
 
-double e_avg_delta() {
-    double ret_val=0;
+float e_avg_delta() {
+    float ret_val=0;
     for (byte i = 0; i < ENGINE_AVG; i++) {
         ret_val += e_deltas[i];
     }
@@ -82,8 +82,8 @@ void s_isr() {
     else s_delta_i++;
 }
 
-double s_avg_delta() {
-    double ret_val=0;
+float s_avg_delta() {
+    float ret_val=0;
     for (byte i = 0; i < SECONDARY_AVG; i++) {
         ret_val += s_deltas[i];
     }
