+++ src/utils.h
@@ -2,7 +2,7 @@
 #define UTILS_H
 
 bool split_string(char** left, char** right, char delim);
-double interpolate(double x, const double *xValues, const double *yValues);
+float interpolate(float x, const float *xValues, const float *yValues);
 bool check_serial();
 bool set(char* token);
 bool zero(char* rest);
